6154 Parameterize Map layers
